LLM_BACKEND/LLM_BACKEND-main/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.4
flask-jwt-extended>=4.0.0
flask-cors==3.0.10
gunicorn>=21
Werkzeug==2.2.3
requests==2.28.1
pymysql==1.1.1
cryptography>=42
pinecone>=5.0.0
flask-socketio
gevent
gevent-websocket